move utility functions from class-decor to functional
